Learnings
1. I wanted the flows to not emit anything if permission was not
   granted, this would simplify the logic in the PlacesAndAssetScreen
composables by not having them know about the permissions when observing
the BleDevicesViewModel.. I wanted to combine two flows  one for the
permissions and one for devices to use a type: `
StateFlow<List<InternalBtDevice>?>`.  Here is the incantation:
```
    val bleDevices = _bleDevices.combine(
        bleRepository.bluetoothPermissionsGranted
    ) { devices, granted -> if (granted) devices else null }.stateIn(
        scope = viewModelScope,
        started = kotlinx.coroutines.flow.SharingStarted.Lazily,
        initialValue = null
    )
```
2. Should I use a sharedFlow or a stateFlow to collect the bleDevices in the Composable.
    val bleDevices = _bleDevices.combine(
        bleRepository.bluetoothPermissionsGranted
    ) { devices, granted -> if (granted) devices else null }.shareIn(
        scope = viewModelScope,
        started = kotlinx.coroutines.flow.SharingStarted.Lazily,
        initialValue = null
    )

The difference is subtle but important. I dono't care about all updates
to the bleDevices list, if they happen in quick succession its ok to get
the final state.

=== State Flow
1. StateFlow maintains a single, current value that is accessible through its value property.
2. When a new value is emitted using the value property or the emit() function, the StateFlow updates its internal state to hold the new value.
3. If multiple values are emitted before a collector has a chance to process them, the StateFlow will only deliver the latest value to the collector.
So for exmaple:
```
val stateFlow = MutableStateFlow(0)

stateFlow.value = 1
stateFlow.value = 2
stateFlow.value = 3

// Collector starts collecting
val collector = launch {
    val value = stateFlow.value
    println("Received: $value")
}
```
The above might miss 1,2 and only print 3

===Shared Flow
If we Look at the case where replay is equal to 0, then the difference
is more subtle because Any *New* Subscribers will not get previously
values. The difference is subtler now, because we can Clean look at just
the values that are limited to the new collector.

1. A shared flow maintains a buffer suspension parameter as is
   documented in the doc string. If a new values are emitted, but the
old ones aren't collected, yet, the flow can suspend the emission and
thus the collector can get all of them.

2. Moreover a Stateflow could have an initial value in the UI code like a
   Null/empty etc, a SharedFlow cannot. You could carefully use
emit(None) from the ViewModel or DB layer initially, but it is harder to maintain.
As an exampled:

```
val sharedFlow = MutableSharedFlow<Int>(replay = 0)

sharedFlow.emit(1)
sharedFlow.emit(2)

// New subscriber joins
val collector = launch {
    sharedFlow.collect { value ->
        println("Received: $value")
    }
}
Here
```